Update button container
