+++ src/pages/home/Home.tsx
@@ -9,7 +9,7 @@ import { DataServices } from '../../shared/services/api/data/DataService';
 import Container from '../../shared/components/Container';
 import Table from '../../shared/components/Table';
 import Line from '../../shared/components/Line/Line';
-import IconsArrow from '../../shared/icons/Arrow/IconArrow';
+import ButtonContainer from '../../shared/components/ButtonContainer/ButtonContainer';
 
 export function Home() {
     const [medicines, setMedicines] = useState<MedicineInterface[]>([]);
@@ -38,28 +38,6 @@ export function Home() {
         });
     }, [page]);
 
-    const handleFirst = () => {
-        if (!pageSettings?.first) return;
-
-        setPage(pageSettings.first);
-    };
-
-    const handleNext = () => {
-        if (page === pageSettings?.pages) return;
-        setPage(page + 1);
-    };
-
-    const handleBack = () => {
-        if (page === 1) return;
-        setPage(page - 1);
-    };
-
-    const handleLast = () => {
-        if (!pageSettings?.last) return;
-
-        setPage(pageSettings.last);
-    };
-
     return (
         <>
             <Container>
@@ -78,44 +56,11 @@ export function Home() {
                         );
                     })}
                 </Table>
-                <div className="buttonMainContainer">
-                    <p>
-                        Total de <b>{pageSettings?.items}</b> item(s)
-                    </p>
-
-                    <div className="buttons">
-                        <p>
-                            Página {page} de {pageSettings?.pages}
-                        </p>
-
-                        <div className="buttonsContainer">
-                            <button
-                                onClick={() => handleFirst()}
-                                className="button first"
-                            >
-                                <IconsArrow name="doubleArrowLeft" />
-                            </button>
-                            <button
-                                onClick={() => handleBack()}
-                                className="button back"
-                            >
-                                <IconsArrow name="arrowLeft" />
-                            </button>
-                            <button
-                                onClick={() => handleNext()}
-                                className="button"
-                            >
-                                <IconsArrow name="arrowRight" />
-                            </button>
-                            <button
-                                onClick={() => handleLast()}
-                                className="button last"
-                            >
-                                <IconsArrow name="doubleArrowRight" />
-                            </button>
-                        </div>
-                    </div>
-                </div>
+                <ButtonContainer
+                    page={page}
+                    pageSettings={pageSettings}
+                    setPage={setPage}
+                />
             </Container>
         </>
     );
