+++ src/shared/icons/Pdf/IconPdf.css
@@ -1,0 +1,3 @@
+svg {
+    fill: black;
+}
